+++ src/app/_components/desktopNavBar.js
@@ -1,6 +1,7 @@
 import { useState, useEffect } from "react";
 import Link from "next/link";
 import Image from "next/image";
+import useAuth from "./hooks/useAuth";
 
 const NAVBAR_HEIGHT = 72;
 const DROPDOWN_BAR_HEIGHT = 256;
@@ -9,57 +10,10 @@ const MENU_BUTTON_HEIGHT = 20;
 const DesktopNavbar = () => {
   const [isMenuDropdownOpen, setIsMenuDropdownOpen] = useState(false);
   const [isUserDropdownOpen, setIsUserDropdownOpen] = useState(false);
-  const [isLoggedIn, setIsLoggedIn] = useState(false); // 로그인 여부 상태
-  const [isLoading, setIsLoading] = useState(true); //로그인 시도중인 상태
-  const [userName, setUserName] = useState(""); // 사용자 이름
 
-  useEffect(() => {
-    fetch(`${process.env.NEXT_PUBLIC_BASE_URL}/auth/profile`, {
-      method: "GET",
-      credentials: "include",
-    })
-      .then((response) => {
-        if (response.status === 200) {
-          return response.json();
-        } else if (response.status === 401) {
-          throw new Error("Unauthorized");
-        } else {
-          throw new Error("Unexpected error");
-        }
-      })
-      .then((data) => {
-        setIsLoggedIn(true);
-        setUserName(data.name);
-      })
-      .catch((error) => {
-        if (error.message === "Unauthorized") {
-          setIsLoggedIn(false);
-        }
-      })
-      .finally(() => {
-        setIsLoading(false);
-      });
-  }, []);
+  const { isLoggedIn, isLoading, userName, redirectToLogin, handleLogout } =
+    useAuth();
 
-  const redirectToLogin = () => {
-    window.location.href =
-      "https://easily-dashboard.jojicompany.com/login?fallback=" +
-      window.location.href;
-  };
-
-  const handleLogout = () => {
-    fetch(`${process.env.NEXT_PUBLIC_BASE_URL}/auth/logout`, {
-      method: "POST",
-      credentials: "include",
-    }).then((response) => {
-      if (response.status === 201) {
-        setIsLoggedIn(false);
-        setIsUserDropdownOpen(false);
-      } else {
-        console.error("Failed to logout.");
-      }
-    });
-  };
   const menuItems = [
     {
       label: "홈",
